services/distributor-nexus/requirements.txt
pika==1.3.2
sqlalchemy==2.0.23
psycopg2==2.9.9
python-dotenv==1.0.0
requests==2.31.0
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
facebook-business==18.0.2
tweepy==4.14.0
linkedin-api==2.0.0
schedule==1.2.0
celery==5.3.4
redis==5.0.1
fastapi==0.104.1
uvicorn==0.24.0
